+++ test/test_collision.py
@@ -153,3 +153,11 @@ class TestCollision(EnvironmentSetup):
             self.env.AddRobot(robot)
             robot.SetDOFValues([-0.91,2.05],[1,3])
             assert(robot.CheckSelfCollision())
+
+    def test_known_collisions(self):
+        env=self.env
+        env.Load('data/lab1.env.xml')
+        robot=env.GetRobots()[0]
+        robot.SetDOFValues([ -8.44575603e-02,   1.48528347e+00,  -5.09108824e-08, 6.48108822e-01,  -4.57571203e-09,  -1.04008750e-08, 7.26855048e-10,   5.50807826e-08,   5.50807826e-08, -1.90689327e-08,   0.00000000e+00])
+        assert(env.CheckCollision(robot))
+