+++ docs/en/changelog.rst
@@ -79,6 +79,8 @@ Planning
 * **Major**: Added :ref:`parabolic smoothing <planner-parabolicsmoother>` as defualt smoother. The parabolic smoothing library is
 from `Kris Hauser's group at Indiana University <http://www.iu.edu/~motion/software.html>`_.
 
+* added jittering of goals to ManipulatorIKGoalSampler
+
 Trajectories
 ------------